+++ web/app/components/workflow/hooks/index.ts
@@ -8,3 +8,4 @@ export * from './use-workflow-run'
 export * from './use-workflow-template'
 export * from './use-checklist'
 export * from './use-workflow-mode'
+export * from './use-workflow-interactions'
